# docs/forge/openapi.yaml
openapi: 3.1.0
info:
  title: Forge — Sandboxed Code Execution
  description: |
    Runs user-submitted commands in isolated containers (Kubernetes or Docker).
    Executions are queued and processed asynchronously by a worker pool.

    ## Authentication

    All endpoints require a Gatekeeper-issued Bearer token:
    `Authorization: Bearer <token>`

    Forge verifies permissions directly with Gatekeeper rather than trusting
    the `X-User-ID` header injected by Conductor, to prevent privilege escalation
    via a compromised gateway.

    ## Allowed images

    The operator configures an explicit allowlist of permitted container images
    via the `ALLOWED_IMAGES` environment variable. Submissions referencing an
    image not in the allowlist are rejected with `400 Bad Request`.

    ## Environment variable restrictions

    The `env` map must use POSIX-compliant key names (`[A-Za-z_][A-Za-z0-9_]*`).
    Keys that could redirect interpreter or linker behaviour (e.g. `LD_PRELOAD`,
    `PYTHONPATH`, `NODE_OPTIONS`) are explicitly blocked.
  version: 1.0.0
  contact:
    email: jhp1403@gmail.com
servers:
  - url: http://localhost:8083
    description: Local development server
security:
  - bearerAuth: []
paths:
...
  /runtime-backends:
    get:
      summary: List runtime backends
      operationId: listRuntimeBackends
      description: |
        Returns all runtime backends ordered by name. Read access is granted to
        users so a client can show which backend a runner class targets; create,
        update, and delete are admin-only.

        Permission: `listRuntimeBackend` on `forge/runtime-backends`.
      responses:
        '200':
          description: List of runtime backends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuntimeBackend'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create runtime backend
      operationId: createRuntimeBackend
      description: |
        Creates a runtime backend (admin only). Returns `409` if one with the same
        name already exists. A `kata` or `gvisor` backend validates that its
        `runtime_class` config key is set.

        Permission: `createRuntimeBackend` on `forge/runtime-backends`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeBackendBody'
      responses:
        '201':
          description: Runtime backend created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBackend'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Runtime backend already exists
          content:
            text/plain:
              schema:
                type: string
  /runtime-backends/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
        description: Runtime backend name

    get:
      summary: Get runtime backend
      operationId: getRuntimeBackend
      description: |
        Permission: `getRuntimeBackend` on `forge/runtime-backends/{name}`.
      responses:
        '200':
          description: Runtime backend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBackend'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update runtime backend
      operationId: updateRuntimeBackend
      description: |
        Replaces the backend's mutable fields (admin only). The registry evicts its
        cached runtime so the change applies to new jobs without a restart.

        Permission: `updateRuntimeBackend` on `forge/runtime-backends/{name}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeBackendBody'
      responses:
        '200':
          description: Updated runtime backend
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBackend'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete runtime backend
      operationId: deleteRuntimeBackend
      description: |
        Deletes the named backend (admin only). The `default` backend cannot be
        deleted and returns `409`.

        Permission: `deleteRuntimeBackend` on `forge/runtime-backends/{name}`.
      responses:
        '204':
          description: Runtime backend deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot delete the default runtime backend
          content:
            text/plain:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token issued by the Gatekeeper service
...
  schemas:
...
    RuntimeBackend:
      type: object
      properties:
        name:
          type: string
          description: Unique backend name. Runner classes reference it by name.
          example: kata-prod
        type:
          type: string
          enum: [docker, kubernetes, kata, gvisor]
          description: |
            Runtime implementation this backend selects. `kata` is the Kubernetes
            runtime pinned to a VM-isolating RuntimeClass; `gvisor` pins a gVisor
            (runsc) RuntimeClass whose userspace kernel sandboxes the job without
            any hardware virtualization.
          example: kata
        enabled:
          type: boolean
          description: Whether the backend may be used by new jobs.
          example: true
        config:
          type: object
          additionalProperties:
            type: string
          description: |
            Non-secret backend settings. For kata: `runtime_class` (required, the
            Kubernetes RuntimeClass, e.g. `kata-qemu`). For gvisor: `runtime_class`
            (required, the gVisor RuntimeClass, e.g. `gvisor`). kubernetes accepts
            an optional `runtime_class` (else K8S_RUNTIME_CLASS). docker reads env
            instead.
          example:
            runtime_class: kata-qemu
        secret_refs:
          type: object
          additionalProperties:
            type: string
          description: |
            Maps a logical key to the NAME of an environment variable read via the
            secret() convention — only env var names are stored and returned, never
            secret values. The built-in backends (docker/kubernetes/kata/gvisor)
            need none.
          example: {}
        created_at:
          type: string
          format: date-time
    RuntimeBackendBody:
      type: object
      required:
        - name
        - type
        - enabled
      properties:
        name:
          type: string
          description: Unique backend name. Ignored on PUT (name comes from path).
          example: kata-prod
        type:
          type: string
          enum: [docker, kubernetes, kata, gvisor]
          example: kata
        enabled:
          type: boolean
          example: true
        config:
          type: object
          additionalProperties:
            type: string
        secret_refs:
          type: object
          additionalProperties:
            type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Missing or invalid Authorization header
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: Insufficient permissions
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Execution not found or not owned by the caller
      content:
        text/plain:
          schema:
            type: string